requirements.txt
fastapi
uvicorn
openai
pypdf
python-multipart
mysql-connector-python